+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -271,6 +271,7 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       'INTEGRATION_TEST_SHOULD_REPORT_RESULTS_TO_NATIVE': 'false',
       'PATROL_TEST_LABEL_ENABLED': displayLabel.toString(),
       'PATROL_TEST_DIRECTORY': config.testDirectory,
+      'PATROL_SCREENSHOT_ON_FAILURE': config.screenshotOnFailure.toString(),
       if (device.targetPlatform != TargetPlatform.web) ...{
         'PATROL_TEST_SERVER_PORT': super.testServerPort.toString(),
         'PATROL_APP_SERVER_PORT': super.appServerPort.toString(),
@@ -326,6 +327,7 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       appServerPort: super.appServerPort,
       testServerPort: super.testServerPort,
       uninstall: uninstall,
+      urlSafeTestNames: config.urlSafeTestNames,
     );
 
     final iosOpts = IOSAppOptions(