+++ packages/patrol_cli/test/crossplatform/app_options_test.dart
@@ -43,6 +43,7 @@ void main() {
             '-Pandroid.injected.androidTest.leaveApksInstalledAfterRun=true',
             '-Papp-server-port=1',
             '-Ptest-server-port=2',
+            '-Purl-safe-test-names=false',
           ]),
         );
       });
@@ -77,6 +78,7 @@ void main() {
             '-Pandroid.injected.androidTest.leaveApksInstalledAfterRun=true',
             '-Papp-server-port=1',
             '-Ptest-server-port=2',
+            '-Purl-safe-test-names=false',
           ]),
         );
       });
@@ -119,6 +121,7 @@ void main() {
             '-Pdart-defines=RU1BSUw9dXNlckBleGFtcGxlLmNvbQ==,UEFTU1dPUkQ9bnk0bmNhdA==,Zm9vPWJhcg==',
             '-Papp-server-port=1',
             '-Ptest-server-port=2',
+            '-Purl-safe-test-names=false',
           ]),
         );
       });
@@ -153,6 +156,7 @@ void main() {
             '-Pdart-defines=RU1BSUw9dXNlckBleGFtcGxlLmNvbQ==,UEFTU1dPUkQ9bnk0bmNhdA==,Zm9vPWJhcg==',
             '-Papp-server-port=1',
             '-Ptest-server-port=2',
+            '-Purl-safe-test-names=false',
           ]),
         );
       });
@@ -188,6 +192,7 @@ void main() {
             '-Pandroid.injected.androidTest.leaveApksInstalledAfterRun=true',
             '-Papp-server-port=1',
             '-Ptest-server-port=2',
+            '-Purl-safe-test-names=false',
           ]),
         );
       });