+++ packages/patrol_cli/test/general/pubspec_reader_test.dart
@@ -83,6 +83,48 @@ patrol:
         expect(reader.read().ios.bundleId, equals('com.example.ExampleApp'));
       });
 
+      test('defaults `screenshot_on_failure` to false when absent', () {
+        fs.file('pubspec.yaml').writeAsStringSync('''
+$_pubspecBase
+patrol:
+  app_name: Example
+''');
+
+        expect(reader.read().screenshotOnFailure, isFalse);
+      });
+
+      test('reads `screenshot_on_failure`', () {
+        fs.file('pubspec.yaml').writeAsStringSync('''
+$_pubspecBase
+patrol:
+  app_name: Example
+  screenshot_on_failure: true
+''');
+
+        expect(reader.read().screenshotOnFailure, isTrue);
+      });
+
+      test('defaults `url_safe_test_names` to false when absent', () {
+        fs.file('pubspec.yaml').writeAsStringSync('''
+$_pubspecBase
+patrol:
+  app_name: Example
+''');
+
+        expect(reader.read().urlSafeTestNames, isFalse);
+      });
+
+      test('reads `url_safe_test_names`', () {
+        fs.file('pubspec.yaml').writeAsStringSync('''
+$_pubspecBase
+patrol:
+  app_name: Example
+  url_safe_test_names: true
+''');
+
+        expect(reader.read().urlSafeTestNames, isTrue);
+      });
+
       test('overrides global values with platform-specific ones', () {
         fs.file('pubspec.yaml').writeAsStringSync('''
 $_pubspecBase